// src/app/model/User.ts
export interface User {
[x: string]: any;
    id: number;
    name: string;
    mobile: number;
    phone: number;
    email: string;
    visit_price: number;
    followup_price: number;
    note: string;
    department_id: number;
    createdAt: string;
    updatedAt: string;
    createdBy: string;
    updatedBy: string;
    active: boolean;
    deleted: boolean;
    }
    
 
    